tests/integration/test_repository.py
#!/usr/bin/env python
# -*- coding: utf-8 -*-
# tests/integration/test_repository.py

# """ Integration tests for the repository module. """

# from src.adapters import generic_repo

# def test_repository_create(session):
#     """Test that an item can be created and added to the repository."""
#     # Instantiate the repository with the session
#     repo = generic_repo.GenericRepository(session)
#     ## ToDo: Replace with your repository class name

#     # Create and add an item to the repository
#     item = "new_item"  ## ToDo: Replace with appropriate item type or object
#     repo.add(item)
#     session.commit()

#     # Verify that the item was added to the database
#     rows = list(session.execute(
#         "SELECT * FROM items WHERE item_name = :name", {"name": item}
#         ## ToDo: Modify query as needed
#     ))

#     # Assert that the item exists in the database
#     assert rows == [(item,)]  ## ToDo: Modify assertion to match result format


# def test_repository_modify(session):
#     """Test that an item can be modified in the repository."""
#     # Instantiate the repository with the session
#     repo = generic_repo.GenericRepository(session)
#     ## ToDo: Replace with your repository class name

#     # Add an initial item to modify
#     original_item = "original_item"
#     ## ToDo: Replace with appropriate item type or object
#     repo.add(original_item)
#     session.commit()

#     # Modify the item (Example: change the item name)
#     modified_item = "modified_item"
#     ## ToDo: Define how the item should be modified
#     repo.modify(original_item, modified_item)
#     ## ToDo: Implement or adjust the modify method in your repository
#     session.commit()

#     # Verify that the item was modified in the database
#     rows = list(session.execute(
#         "SELECT * FROM items WHERE item_name = :name",
#         {"name": modified_item}  ## ToDo: Modify query as needed
#     ))

#     # Assert that the modified item exists in the database
#     assert rows == [(modified_item,)]  ## ToDo: Modify assertion to match result


# def test_repository_save(session):
#     """Test that an item can be saved (created or updated) in the repository."""
#     # Instantiate the repository with the session
#     repo = generic_repo.GenericRepository(session)
#     ## ToDo: Replace with your repository class name

#     # Scenario 1: Save a new item
#     item = "item_to_save"  ## ToDo: Replace with appropriate item type or object
#     repo.save(item)  # Assuming save either adds or updates
#     session.commit()

#     # Verify that the item was saved (created) in the database
#     rows = list(session.execute(
#         "SELECT * FROM items WHERE item_name = :name", {"name": item}
#         ## ToDo: Modify query as needed
#     ))
#     assert rows == [(item,)]  ## ToDo: Modify assertion to match result format

#     # Scenario 2: Modify the item and save again
#     modified_item = "saved_modified_item"
#     ## ToDo: Define the modified version of the item
#     repo.save(modified_item)  # Save should update the existing item
#     session.commit()

#     # Verify that the item was updated in the database
#     rows = list(session.execute(
#         "SELECT * FROM items WHERE item_name = :name",
#         {"name": modified_item}  ## ToDo: Modify query as needed
#     ))
#     assert rows == [(modified_item,)]  ## ToDo: Modify assertion to match result
